Add javadoc comments for classes and methods
