+++ src/main/java/Command.java
@@ -1,3 +1,7 @@
+/**
+ * Represents possible command types for Buddy chatbot
+ *
+ */
 public enum Command {
     TODO, DEADLINE, EVENT, BYE,
     LIST, MARK, UNMARK, DELETE
