+++ src/main/java/task/Deadline.java
@@ -1,5 +1,9 @@
 package task;
 
+/**
+ * Represents Deadline event
+ *
+ */
 public class Deadline extends Task {
     protected String by;
