+++ src/main/java/task/Event.java
@@ -1,5 +1,9 @@
 package task;
 
+/**
+ * Represents Event type
+ *
+ */
 public class Event extends Task{
     protected String from;
     protected String to;
